close bug issue 165, re-spotted by Leo Baltus
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns/pdns@1098 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ rec_channel.cc
@@ -13,6 +13,7 @@ using namespace std;
 RecursorControlChannel::RecursorControlChannel()
 {
   d_fd=-1;
+  d_dontclose=false;
   *d_local.sun_path=0;
 }
 
@@ -20,7 +21,7 @@ RecursorControlChannel::~RecursorControlChannel()
 {
   if(d_fd > 0)
     close(d_fd);
-  if(*d_local.sun_path)
+  if(*d_local.sun_path && !d_dontclose)
     unlink(d_local.sun_path);
 }
